PostgreSQL doesn't like INTEGER UNSIGNED

+++ synapse/storage/data_stores/main/schema/delta/56/room_key_etag.sql
@@ -14,4 +14,4 @@
  */
 
 -- store the current etag of backup version
-ALTER TABLE e2e_room_keys_versions ADD COLUMN etag INTEGER UNSIGNED;
+ALTER TABLE e2e_room_keys_versions ADD COLUMN etag BIGINT UNSIGNED;
